changed table generation and data load from scripts (dev)

+++ src/main/java/home/lflt/model/Portfolio.java
@@ -29,7 +29,7 @@ public class Portfolio {
     private String info;
     private String type;
 
-    private int funds;
+    private double funds;
     private int epochs;
 
 //    @CreationTimestamp
